+++ src/packages/array_basic.rs
@@ -235,52 +235,15 @@ pub mod array_functions {
 
         // Check if array will be over max size limit
         #[cfg(not(feature = "unchecked"))]
-        {
-            use crate::types::dynamic::Union;
+        if _ctx.engine().max_array_size() > 0 {
+            let pad = len - array.len();
+            let (a, m, s) = Dynamic::calc_array_sizes(array, true);
+            let (ax, mx, sx) = item.calc_data_sizes(true);
 
-            if _ctx.engine().max_array_size() > 0 && len > _ctx.engine().max_array_size() {
-                return Err(
-                    ERR::ErrorDataTooLarge("Size of array".to_string(), Position::NONE).into(),
-                );
-            }
-
-            let check_sizes = match item.0 {
-                Union::Str(..) => true,
-                Union::Array(..) => true,
-                #[cfg(not(feature = "no_object"))]
-                Union::Map(..) => true,
-                _ => false,
-            };
-
-            if check_sizes {
-                let mut arr_len = array.len();
-                let mut arr = Dynamic::from_array(mem::take(array));
-
-                let (mut a1, mut m1, mut s1) = arr.calc_data_sizes(true);
-                let (a2, m2, s2) = item.calc_data_sizes(true);
-
-                {
-                    let mut guard = arr.write_lock::<Array>().unwrap();
-
-                    while arr_len < len {
-                        a1 += a2;
-                        m1 += m2;
-                        s1 += s2;
-
-                        _ctx.engine().throw_on_size((a1, m1, s1))?;
-
-                        guard.push(item.clone());
-                        arr_len += 1;
-                    }
-                }
-
-                *array = arr.into_array().unwrap();
-            } else {
-                array.resize(len, item);
-            }
+            _ctx.engine()
+                .throw_on_size((a + pad + ax * pad, m + mx * pad, s + sx * pad))?;
         }
 
-        #[cfg(feature = "unchecked")]
         array.resize(len, item);
 
         Ok(())
@@ -644,6 +607,43 @@ pub mod array_functions {
             result
         }
     }
+
+    /// Iterate through all the elements in the array, applying a `process` function to each element in turn.
+    /// Each element is bound to `this` before calling the function.
+    ///
+    /// # Function Parameters
+    ///
+    /// * `this`: bound to array element (mutable)
+    /// * `index` _(optional)_: current index in the array
+    ///
+    /// # Example
+    ///
+    /// ```rhai
+    /// let x = [1, 2, 3, 4, 5];
+    ///
+    /// x.for_each(|| this *= this);
+    ///
+    /// print(x);       // prints "[1, 4, 9, 16, 25]"
+    ///
+    /// x.for_each(|i| this *= i);
+    ///
+    /// print(x);       // prints "[0, 2, 6, 12, 20]"
+    /// ```
+    #[rhai_fn(return_raw)]
+    pub fn for_each(ctx: NativeCallContext, array: &mut Array, map: FnPtr) -> RhaiResultOf<()> {
+        if array.is_empty() {
+            return Ok(());
+        }
+
+        for (i, item) in array.iter_mut().enumerate() {
+            let ex = [(i as INT).into()];
+
+            let _ = map.call_raw_with_extra_args("map", &ctx, Some(item), [], ex, None)?;
+        }
+
+        Ok(())
+    }
+
     /// Iterate through all the elements in the array, applying a `mapper` function to each element
     /// in turn, and return the results as a new array.
     ///
@@ -679,8 +679,7 @@ pub mod array_functions {
 
         for (i, item) in array.iter_mut().enumerate() {
             let ex = [(i as INT).into()];
-
-            ar.push(map.call_raw_with_extra_args("map", &ctx, Some(item), [], ex)?);
+            ar.push(map.call_raw_with_extra_args("map", &ctx, Some(item), [], ex, Some(0))?);
         }
 
         Ok(ar)
@@ -723,7 +722,7 @@ pub mod array_functions {
             let ex = [(i as INT).into()];
 
             if filter
-                .call_raw_with_extra_args("filter", &ctx, Some(item), [], ex)?
+                .call_raw_with_extra_args("filter", &ctx, Some(item), [], ex, Some(0))?
                 .as_bool()
                 .unwrap_or(false)
             {
@@ -962,7 +961,7 @@ pub mod array_functions {
             let ex = [(i as INT).into()];
 
             if filter
-                .call_raw_with_extra_args("index_of", &ctx, Some(item), [], ex)?
+                .call_raw_with_extra_args("index_of", &ctx, Some(item), [], ex, Some(0))?
                 .as_bool()
                 .unwrap_or(false)
             {
@@ -1128,7 +1127,8 @@ pub mod array_functions {
         for (i, item) in array.iter_mut().enumerate().skip(start) {
             let ex = [(i as INT).into()];
 
-            let value = filter.call_raw_with_extra_args("find_map", &ctx, Some(item), [], ex)?;
+            let value =
+                filter.call_raw_with_extra_args("find_map", &ctx, Some(item), [], ex, Some(0))?;
 
             if !value.is_unit() {
                 return Ok(value);
@@ -1169,7 +1169,7 @@ pub mod array_functions {
             let ex = [(i as INT).into()];
 
             if filter
-                .call_raw_with_extra_args("some", &ctx, Some(item), [], ex)?
+                .call_raw_with_extra_args("some", &ctx, Some(item), [], ex, Some(0))?
                 .as_bool()
                 .unwrap_or(false)
             {
@@ -1211,7 +1211,7 @@ pub mod array_functions {
             let ex = [(i as INT).into()];
 
             if !filter
-                .call_raw_with_extra_args("all", &ctx, Some(item), [], ex)?
+                .call_raw_with_extra_args("all", &ctx, Some(item), [], ex, Some(0))?
                 .as_bool()
                 .unwrap_or(false)
             {
@@ -1334,11 +1334,11 @@ pub mod array_functions {
         }
 
         array
-            .iter()
+            .iter_mut()
             .enumerate()
             .try_fold(initial, |result, (i, item)| {
                 let ex = [(i as INT).into()];
-                reducer.call_raw_with_extra_args("reduce", &ctx, None, [result, item.clone()], ex)
+                reducer.call_raw_with_extra_args("reduce", &ctx, Some(item), [result], ex, Some(1))
             })
     }
     /// Reduce an array by iterating through all elements, in _reverse_ order,
@@ -1400,19 +1400,22 @@ pub mod array_functions {
             return Ok(initial);
         }
 
+        let len = array.len();
+
         array
-            .iter()
+            .iter_mut()
             .rev()
             .enumerate()
             .try_fold(initial, |result, (i, item)| {
-                let ex = [((array.len() - 1 - i) as INT).into()];
+                let ex = [((len - 1 - i) as INT).into()];
 
                 reducer.call_raw_with_extra_args(
                     "reduce_rev",
                     &ctx,
-                    None,
-                    [result, item.clone()],
+                    Some(item),
+                    [result],
                     ex,
+                    Some(1),
                 )
             })
     }
@@ -1602,7 +1605,7 @@ pub mod array_functions {
             let ex = [(i as INT).into()];
 
             if filter
-                .call_raw_with_extra_args("drain", &ctx, Some(&mut array[x]), [], ex)?
+                .call_raw_with_extra_args("drain", &ctx, Some(&mut array[x]), [], ex, Some(0))?
                 .as_bool()
                 .unwrap_or(false)
             {
@@ -1749,7 +1752,7 @@ pub mod array_functions {
             let ex = [(i as INT).into()];
 
             if filter
-                .call_raw_with_extra_args("retain", &ctx, Some(&mut array[x]), [], ex)?
+                .call_raw_with_extra_args("retain", &ctx, Some(&mut array[x]), [], ex, Some(0))?
                 .as_bool()
                 .unwrap_or(false)
             {